WEB-4654 connect summary period filter
