+++ app/pages/clinicworkspace/ClinicPatients.js
@@ -51,6 +51,7 @@ import { useFlags, useLDClient } from 'launchdarkly-react-client-sdk';
 import { Link as RouterLink } from 'react-router-dom';
 import useClinicPatientsFilters, { defaultFilterState, SPECIAL_FILTER_STATES } from './useClinicPatientsFilters';
 import AppliedFiltersAdapter from './AppliedFiltersAdapter';
+import FilterBySummaryPeriod from './FilterBySummaryPeriod';
 
 import {
   bindPopover,
@@ -727,24 +728,6 @@ export const ClinicPatients = (props) => {
 
   const customLastDataFilterOptions = reject(lastDataFilterOptions, { value: 7 });
 
-  const summaryPeriodOptions = [
-    { value: '1d', label: t('24 hours') },
-    { value: '7d', label: t('7 days') },
-    { value: '14d', label: t('14 days') },
-    { value: '30d', label: t('30 days') },
-  ];
-
-  const getSummaryPeriodSelectLabel = (activeSummaryPeriod) => {
-    switch(activeSummaryPeriod) {
-      case '1d': return t('Summarizing 24 hours of data');
-      case '7d': return t('Summarizing 7 days of data');
-      case '14d': return t('Summarizing 14 days of data');
-      case '30d': return t('Summarizing 30 days of data');
-    }
-
-    return null;
-  };
-
   const clinicSites = useMemo(() => keyBy(clinic?.sites, 'id'), [clinic?.sites]);
   const patientTags = useMemo(() => keyBy(clinic?.patientTags, 'id'), [clinic?.patientTags]);
 
@@ -762,13 +745,7 @@ export const ClinicPatients = (props) => {
 
   const defaultSummaryPeriod = '14d';
   const [activeSummaryPeriod, setActiveSummaryPeriod] = useLocalStorage('activePatientSummaryPeriod', defaultSummaryPeriod);
-  const [pendingSummaryPeriod, setPendingSummaryPeriod] = useState(activeSummaryPeriod);
   const previousSummaryPeriod = usePrevious(activeSummaryPeriod);
-
-  const summaryPeriodPopupFilterState = usePopupState({
-    variant: 'popover',
-    popupId: 'summaryPeriodFilters',
-  });
 
   const lastDataPopupFilterState = usePopupState({
     variant: 'popover',
@@ -2507,83 +2484,10 @@ export const ClinicPatients = (props) => {
                   sx={{ gap: 3, justifyContent: 'flex-end', alignItems: 'center', flexShrink: 0 }}
                 >
 
-                <Box
-                  onClick={() => {
-                    if (!summaryPeriodPopupFilterState.isOpen) trackMetric(prefixPopHealthMetric('Summary period filter open'), { clinicId: selectedClinicId });
-                  }}
-                >
-                  <Button
-                    variant="filter"
-                    selected={true}
-                    id="summary-period-filter-trigger"
-                    {...bindTrigger(summaryPeriodPopupFilterState)}
-                    icon={KeyboardArrowDownRoundedIcon}
-                    iconLabel="Filter by summary period duration"
-                    sx={{ fontSize: 0, lineHeight: 1.3 }}
-                  >
-                    {getSummaryPeriodSelectLabel(activeSummaryPeriod)}
-                  </Button>
-                </Box>
-
-                <Popover
-                  width="13em"
-                  closeIcon
-                  {...bindPopover(summaryPeriodPopupFilterState)}
-                  onClickCloseIcon={() => {
-                    trackMetric(prefixPopHealthMetric('Summary period filter close'), { clinicId: selectedClinicId });
-                  }}
-                  onClose={() => {
-                    summaryPeriodPopupFilterState.close();
-                    setPendingSummaryPeriod(activeSummaryPeriod);
-                  }}
-                >
-                  <DialogContent px={2} py={3} dividers>
-                    <Body0 color="grays.4" sx={{ fontWeight: 'medium' }} mb={2}>{t('Tidepool will generate health summaries for the selected number of days.')}</Body0>
-
-                    <RadioGroup
-                      id="summary-period-filters"
-                      name="summary-period-filters"
-                      options={summaryPeriodOptions}
-                      variant="vertical"
-                      sx={{ fontSize: 0 }}
-                      value={pendingSummaryPeriod || activeSummaryPeriod}
-                      onChange={event => setPendingSummaryPeriod(event.target.value)}
-                    />
-                  </DialogContent>
-
-                  <DialogActions sx={{ justifyContent: 'space-between' }} p={1}>
-                    <Button
-                      id="cancel-summary-period-filter"
-                      sx={{ fontSize: 1 }}
-                      variant="textSecondary"
-                      onClick={() => {
-                        trackMetric(prefixPopHealthMetric('Summary period filter cancel'), { clinicId: selectedClinicId });
-                        setPendingSummaryPeriod(activeSummaryPeriod);
-                        summaryPeriodPopupFilterState.close();
-                      }}
-                    >
-                      {t('Cancel')}
-                    </Button>
-
-                    <Button
-                      id="apply-summary-period-filter"
-                      sx={{ fontSize: 1 }}
-                      variant="textPrimary"
-                      disabled={pendingSummaryPeriod === activeSummaryPeriod}
-                      onClick={() => {
-                        trackMetric(prefixPopHealthMetric('Summary period apply filter'), {
-                          clinicId: selectedClinicId,
-                          summaryPeriod: pendingSummaryPeriod,
-                        });
-
-                        setActiveSummaryPeriod(pendingSummaryPeriod);
-                        summaryPeriodPopupFilterState.close();
-                      }}
-                    >
-                      {t('Apply')}
-                    </Button>
-                  </DialogActions>
-                </Popover>
+                <FilterBySummaryPeriod
+                  activeSummaryPeriod={activeSummaryPeriod}
+                  setActiveSummaryPeriod={setActiveSummaryPeriod}
+                />
 
                 {showRpmReportUI && (
                   <Flex
